Move code out and make it loadable in browser(s) too.
